+++ CHANGES.md
@@ -27,6 +27,7 @@ Changes in 4.4-beta
 - Add PasswordStrength jQuery plugin in jquery-passwordstrength.js & plugins.min.js
 - Use PasswordInput() function in both General_Info.inc.php
 - JS Add zxcvbn (password strength estimator) in zxcvbn.js
+- Add PASSWORD_STRENGTH to CONFIG table in Update.fnc.php & rosariosis.sql
 - CSS bigger input (checkbox, select) size in stylesheet.css
 - JS Load once on page load & always check height on resize & scroll in jquery-fixedmenu.js
 - JS Open submenu on touch (mobile & tablet) in warehouse.js